third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/push_messaging/subscribe-success-in-document-expected.txt

This is a testharness.js-based test.
PASS subscribe() succeeds when permission is granted and resolves with a valid subscription 
Harness: the test ran to completion.
